Support cancellation propogation
To support cancellation on the server we introduce a map tracking all pending
calls and an internal service which is registered on all servers. The pending
map is similiar to what is done on the client, except instead of tracking
pending Calls, we just track context cancel function for pending requests. The
internal service cancel method takes the requests sequence number and the
pending map to call cancel. We have to special case the cancel method in Call
so that we can pass along the pending map.