Unexport Pending field from CancelArgs
The intention of this is to make the cancel message smaller on the wire.
Previously `encoding/gob` would encode the empty `Pending` field onto the wire
for each cancel RPC. This was an oversight. Luckily we can just unexport it and
since the field was unused this change is both backwards and forwards
compatible.

Tested cancellations with the combinations `{old_client,
new_client}x{old_server, new_server}` using
https://gist.github.com/keegancsmith/a53fbd2d4849343d6ca70b27fd062667