introduce ClientTrace
This introduces a ClientTrace. This is based off
net/http/httptrace.ClientTrace. We introduce it in a way were we keep
backwards compatibility. This is useful for instrumenting rpc to
understand time taken to acquire the request mutex, as well as time
taken to marshal and write the request down the wire.

We require the context to get the ClientTrace value. As such we
introduce a new function Client.GoContext. This is Client.Go, but with a
context argument.